tool: move serial command under time
The serial command needs access to time/internal/serialpps for its PPS monitoring mode, but its repository-root internal location cannot import that package under Go internal visibility rules.

Relocate serialcmd to time/app alongside the daemon command layer, and update the satpulsetool import and package documentation.
