+++ cmd/satpulsetool/commands.go
@@ -12,7 +12,7 @@ import (
 	"github.com/jclark/satpulse/internal/packcmd"
 	"github.com/jclark/satpulse/internal/replaycmd"
 	"github.com/jclark/satpulse/internal/scancmd"
-	"github.com/jclark/satpulse/internal/serialcmd"
+	"github.com/jclark/satpulse/time/app/serialcmd"
 )
 
 type cmdFunc func(logWriter io.Writer, logLevel slog.Level, progName string, cmdName string, cmdArgs []string) (usage string, err error)
